+++ pages/_app.tsx
@@ -1,96 +1,9 @@
 import '../styles/global.css';
 import { FunctionComponent, useState } from "react";
-
-type Mark = 'X' | 'O' | null;
-type SquareProps = {
-  value: Mark;
-  onSquareClick: () => void;
-}
-
-type BoardProps = {
-  xIsNext: boolean;
-  squares: Array<Mark>;
-  onPlay: (nextSquares: Array<Mark>) => void;
-}
-
-const Square: FunctionComponent<SquareProps> = ({ value, onSquareClick }: SquareProps) => {
-  return (
-    <button className="square" onClick={onSquareClick}>
-      {value}
-    </button>
-  );
-}
-
-const Board: FunctionComponent<BoardProps> = ({ xIsNext, squares, onPlay }: BoardProps) => {
-  const calculateWinner = (squares: Array<Mark>) => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (
-        squares[a] &&
-        squares[a] === squares[b] &&
-        squares[a] === squares[c]
-      ) {
-        return squares[a];
-      }
-    }
-    return null;
-  }
-
-  const handleClick = (i: number) => {
-    if (calculateWinner(squares) || squares[i]) {
-      return;
-    }
-    const nextSquares: Array<Mark> = squares.slice();
-    if (xIsNext) {
-      nextSquares[i] = "X";
-    } else {
-      nextSquares[i] = "O";
-    }
-    onPlay(nextSquares);
-  }
-
-  const winner = calculateWinner(squares);
-  let status;
-  if (winner) {
-    status = "Winner: " + winner;
-  } else {
-    status = "Next player: " + (xIsNext ? "X" : "O");
-  }
-
-  return (
-    <>
-      <div className="status">{status}</div>
-      <div className="board-row">
-        <Square value={squares[0]} onSquareClick={() => handleClick(0)} />
-        <Square value={squares[1]} onSquareClick={() => handleClick(1)} />
-        <Square value={squares[2]} onSquareClick={() => handleClick(2)} />
-      </div>
-      <div className="board-row">
-        <Square value={squares[3]} onSquareClick={() => handleClick(3)} />
-        <Square value={squares[4]} onSquareClick={() => handleClick(4)} />
-        <Square value={squares[5]} onSquareClick={() => handleClick(5)} />
-      </div>
-      <div className="board-row">
-        <Square value={squares[6]} onSquareClick={() => handleClick(6)} />
-        <Square value={squares[7]} onSquareClick={() => handleClick(7)} />
-        <Square value={squares[8]} onSquareClick={() => handleClick(8)} />
-      </div>
-    </>
-  );
-}
+import { Mark } from "../types/index";
+import { Board } from "../components/board";
 
 export default function Game() {
-  // 9個の要素を持つ配列をnullで初期化
   const [history, setHistory] = useState([Array(9).fill(null)]);
   const [currentMove, setCurrentMove] = useState(0);
   const xIsNext = currentMove % 2 === 0;
